\name{Catheter}
\alias{Catheter }
\docType{data}
\title{The Catheter Segment Culture data.}
\description{
  The studies studied semi-quantitative (19 studies) and quantitative (14 studies) catheter segment culture for the diagnosis of intravascular device-related blood stream infection.
}
\usage{data(Catheter)}
\format{
  A data frame with 33 rows and 8 columns. Each row provides
  data for one study.
}
%%\details{}
\references{Chu H, Guo H, Zhou Y. Bivariate random effects meta-analysis of diagnostic studies using generalized linear mixed models[J]. Medical Decision Making, 2009.}

\examples{
  data(Catheter)
}
\keyword{datasets}