\name{meta4diagGUI}
\alias{meta4diagGUI}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
	A graphic user interface for meta4diag.
}
\description{
	A graphic user interface for meta4diag which is fully interactive but still provides full functionality.
}
\usage{
	meta4diagGUI()
}

\references{
  Havard Rue, Sara Martino, and Nicholas Chopin (2009). Approximate
  Bayesian Inference for Latent Gaussian Models Using Integrated Nested
  Laplace Approximations. Journal of the Royal Statistical Society B,
  71, 319-392. (www.r-inla.org)

  
}
\author{
  Jingyi Guo and Andrea Riebler 
}
\seealso{\code{\link{meta4diag}}}
\examples{
\dontrun{meta4diagGUI()}
}