\name{Telomerase}
\alias{Telomerase}
\docType{data}
\title{Telomerase dataset.}
\description{
  The dataset contains 10 studies from a meta-analysis that uses the telomerase marker for the diagnosis of bladder cancer.
}
\usage{data(Telomerase)}
\format{
  A data frame with 10 rows and 5 columns. Each row provides data for one study. The columns are "studynames" indicates the name of studies, "TP", "FP", "FN" and "TN" which are the number of true positive, false positive, false negative and true negative, respectively.
}
%%\details{}
\references{
Glas AS, Roos D, Deutekom M, Zwinderman AH, Bossuyt PM, Kurth KH (2003). "Tumor Markers in the Diagnosis of Primary Bladder Cancer. A Systematic Review." The Journal of Urology, 169(6), 1975-1982.
}

\examples{
  data(Telomerase)
}
\keyword{datasets}